Make details responsive
